# docs/swagger.yaml
basePath: /
definitions:
...
  ConfigurableResourceSummary:
    properties:
      description:
        type: string
      display_name:
        type: string
      key:
        type: string
    type: object
...
  JSON:
    additionalProperties: {}
    type: object
...
  connectionResponse:
    properties:
      actions_count:
        type: integer
      configurable_resources:
        items:
          $ref: '#/definitions/ConfigurableResourceSummary'
        type: array
      created_at:
        type: string
      display_name:
        type: string
      id:
        type: string
      integration_id:
        type: string
      meta:
        $ref: '#/definitions/JSON'
      nango_connection_id:
        type: string
      org_id:
        type: string
      provider:
        type: string
      provider_config:
        $ref: '#/definitions/JSON'
      revoked_at:
        type: string
      updated_at:
        type: string
      webhook_configured:
        type: boolean
    type: object
...
  createConnectionRequest:
    properties:
      meta:
        $ref: '#/definitions/JSON'
      nango_connection_id:
        type: string
    type: object
...
  errorResponse:
    properties:
      error:
        type: string
    type: object
...
host: api.dev.usehivy.com
info:
  contact: {}
  description: Proxy runtime for LLM API credentials.
  title: Hivy API
  version: "1.0"
paths:
...
  /incoming/gateways/external/{routeID}:
    post:
      consumes:
      - application/json
      parameters:
      - description: Gateway route ID
        in: path
        name: routeID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
      summary: Receive an external gateway message
      tags:
      - webhooks
...
  /v1/connections/{id}:
    delete:
      description: Revokes a user's platform integration connection and removes it
        from Nango.
      parameters:
      - description: Connection ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorResponse'
      security:
      - BearerAuth: []
      summary: Disconnect an connection
      tags:
      - connections
...
  /v1/integrations/{id}/connections:
    post:
      consumes:
      - application/json
      description: Stores a connection after the OAuth flow completes via Nango.
      parameters:
      - description: Integration ID
        in: path
        name: id
        required: true
        type: string
      - description: Connection details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/createConnectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/connectionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorResponse'
      security:
      - BearerAuth: []
      summary: Create an connection
      tags:
      - connections
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Bearer token (JWT or API key).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"